mama_bear_backend/mama_bear_proj/requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
pillow==10.4.0
psycopg==3.2.1
psycopg-binary==3.2.1
sqlparse==0.5.1
typing_extensions==4.12.2
